+++ Allgemeines/openweathermap/jsonparser/jsonparser/currentweather/Main.cs
@@ -5,8 +5,8 @@ namespace jsonparser.currentweather
     public class Main
     {
         public double? temp { get; set; }
-        public int? pressure { get; set; }
-        public int? humidity { get; set; }
+        public double? pressure { get; set; }
+        public double? humidity { get; set; }
         public double? temp_min { get; set; }
         public double? temp_max { get; set; }
     }
