+++ Allgemeines/openweathermap/jsonparser/jsonparser/currentweather/Sys.cs
@@ -6,8 +6,8 @@ namespace jsonparser.currentweather
     {
         public double? message { get; set; }
         public string country { get; set; }
-        public int? sunrise { get; set; }
-        public int? sunset { get; set; }
+        public long? sunrise { get; set; }
+        public long? sunset { get; set; }
     }
 }
