+++ Allgemeines/openweathermap/jsonparser/jsonparser/forecastweather/City.cs
@@ -4,11 +4,11 @@ namespace jsonparser.forecastweather
 {
     public class City
     {
-        public int? id { get; set; }
+        public long? id { get; set; }
         public string name { get; set; }
         public Coord coord { get; set; }
         public string country { get; set; }
-        public int? population { get; set; }
+        public long? population { get; set; }
     }
 }
